Fixed iPad issue 6: More timers on start need to be shown
Prior to starting timers, or when user re-opens app, only one timer is
visible (unless user scrolls). More timers need to be visible - up to
five. If necessary, move the start button down to center the list on
the circle to the left. Basically, move the list up so more timers can
be seen. In the screen shot, there are 5 timers but user can only see 1
unless user scrolls.